+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 - Improve Tooltip performance
 
+### Deprecated
+
+- Deprecate `closest()` and `within()` functions
+
+### Fixed
+
+- Fix Drop and Modal based components remain open when anchor click targets element within
+
 ## 3.17.8 (November 2, 2023)
 
 ### Fixed
@@ -32,8 +40,7 @@
 
 ### Deprecated
 
-- Deprecate `isInView()` function
-- Deprecate `intersectRect()` function
+- Deprecate `isInView()` and `intersectRect()` functions
 - Deprecate `target-offset` option in Sticky component
 
 ### Changed